+++ src/challenge/lv1/Input.java
@@ -1,0 +1,4 @@
+package challenge.lv1;
+
+public class Input {
+}
