+++ src/lv2/KioskMain2.java
@@ -1,4 +1,4 @@
-package Lv2;
+package lv2;
 
 import java.util.ArrayList;
 import java.util.List;
