+++ src/lv3/Kiosk3.java
@@ -1,4 +1,4 @@
-package Lv3;
+package lv3;
 
 import java.util.ArrayList;
 import java.util.List;
