+++ src/lv3/MenuItem3.java
@@ -1,4 +1,4 @@
-package Lv3;
+package lv3;
 
 public class MenuItem3 {
     private final int numbering;
