+++ src/lv4lv5/Menu.java
@@ -1,4 +1,4 @@
-package Lv4Lv5;
+package lv4lv5;
 
 import java.util.ArrayList;
 import java.util.HashMap;
