+++ src/lv4lv5/MenuItem4.java
@@ -1,17 +1,23 @@
-package Lv2;
+package lv4lv5;
 
-public class MenuItem2 {
-    private String name;
-    private double price;
-    private String description;
+public class MenuItem4 {
+    private final int numbering;
+    private final String name;
+    private final double price;
+    private final String description;
 
-    public MenuItem2(String name, double price, String description){
+    public MenuItem4(int numbering, String name, double price, String description){
+        this.numbering = numbering;
         this.name = name;
         this.price = price;
         this.description = description;
     }
 
     //getter
+    //메뉴 번호 가져오기'
+    public int getNumbering(){
+        return numbering;
+    }
     //햄버거 이름 가져오기
     public String getName(){
         return name;
